refactor(#5): extract overview column, sheet, and share helpers
Thin dashboard-overview by moving column persistence, detail-sheet state, and share-link copy flows into dedicated modules.
